requirements.txt
numpy==1.19.5
Flask==2.1.0
keras==2.7.0
Pillow==8.4.0
gunicorn==20.1.0
sklearn
matplotlib
tensorflow-cpu
opencv-python-headless==4.5.3.56
scipy
pandas
scikit-learn
